# Improve speed when we have many candidate matches